# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hängt! – Getränkeabrechnungs-API
  version: 0.1.0
  description: |
    REST-API für die vereinsinterne Getränkeabrechnung.
    Alle geschützten Endpunkte erwarten einen JWT im `Authorization`-Header.
servers:
  - url: /api/v1
    description: Lokaler Dev-Server (http://localhost:3001/api/v1)

# ---------------------------------------------------------------------------
# Sicherheit
# ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # Wiederverwendbare Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Mitglied nicht gefunden
        code:
          type: string
          example: NOT_FOUND
        details:
          type: object
          description: Zod-Flatten-Ausgabe bei Validierungsfehlern

    # -- Member ----------------------------------------------------------------
    Member:
      type: object
      required: [id, username, display_name, role, is_active, created_at, updated_at]
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: max.mustermann
        display_name:
          type: string
          example: Max Mustermann
        role:
          type: string
          enum: [admin, member]
          example: member
        is_active:
          type: integer
          enum: [0, 1]
          example: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateMemberRequest:
      type: object
      required: [username, display_name, password]
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 50
          pattern: '^[a-zA-Z0-9._-]+$'
          example: max.mustermann
        display_name:
          type: string
          minLength: 1
          maxLength: 100
          example: Max Mustermann
        password:
          type: string
          minLength: 8
          maxLength: 72
          example: sicheresPasswort1
        role:
          type: string
          enum: [admin, member]
          default: member
    UpdateMemberRequest:
      type: object
      minProperties: 1
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 100
          example: Max Mustermann
        password:
          type: string
          minLength: 8
          maxLength: 72
        role:
          type: string
          enum: [admin, member]

    # -- Drink -----------------------------------------------------------------
...
  responses:
    Unauthorized:
      description: Kein oder ungültiges Token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Kein Authentifizierungstoken angegeben
    Forbidden:
      description: Nicht ausreichende Berechtigung
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unzureichende Berechtigung
    NotFound:
      description: Ressource nicht gefunden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Mitglied nicht gefunden
            code: NOT_FOUND
    BadRequest:
      description: Ungültige Eingabe (Validierungsfehler)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

# ---------------------------------------------------------------------------
# Endpunkte
# ---------------------------------------------------------------------------
security:
  - bearerAuth: []
paths:
...
  /members:
    get:
      tags: [Members]
      summary: Alle Mitglieder
      description: Gibt alle Mitglieder zurück. Nur für Admins.
      parameters:
        - name: includeInactive
          in: query
          schema:
            type: boolean
            default: false
          description: Wenn true, werden auch deaktivierte Mitglieder zurückgegeben.
      responses:
        '200':
          description: Liste der Mitglieder (ohne password_hash)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Member'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Members]
      summary: Mitglied anlegen
      description: Legt ein neues Vereinsmitglied an. Nur für Admins.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMemberRequest'
      responses:
        '201':
          description: Mitglied angelegt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Username bereits vergeben
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Username bereits vergeben
                code: USERNAME_TAKEN
  /members/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1

    get:
      tags: [Members]
      summary: Einzelnes Mitglied
      responses:
        '200':
          description: Mitglied gefunden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Members]
      summary: Mitglied aktualisieren
      description: Aktualisiert Display-Name, Passwort und/oder Rolle. Mindestens ein Feld erforderlich.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberRequest'
      responses:
        '200':
          description: Mitglied aktualisiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Members]
      summary: Mitglied deaktivieren (Soft-Delete)
      description: |
        Setzt `is_active = 0`. Das Mitglied kann sich danach nicht mehr einloggen.
        Ein Admin kann sich nicht selbst deaktivieren.
      responses:
        '204':
          description: Mitglied deaktiviert
        '400':
          description: Selbst-Deaktivierung nicht erlaubt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Eigenes Konto kann nicht deaktiviert werden
                code: SELF_DEACTIVATION
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Mitglied ist bereits deaktiviert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Mitglied ist bereits deaktiviert
                code: ALREADY_INACTIVE

  # ==========================================================================
  # Drinks
  # ==========================================================================
...